Move Gallery to base (gallery) module to domain package
