+++ app/src/main/java/com/ataulm/artcollector/domain/Gallery.kt
@@ -1,0 +1,3 @@
+package com.ataulm.artcollector.domain
+
+class Gallery(collection: Collection<Painting>) : ArrayList<Painting>(collection)
