+++ app/src/main/java/com/ataulm/artcollector/gallery/domain/GalleryRepository.kt
@@ -1,5 +1,7 @@
 package com.ataulm.artcollector.gallery.domain
 
+import com.ataulm.artcollector.domain.Gallery
+
 internal interface GalleryRepository {
 
     suspend fun gallery(): Gallery
