+++ artist/src/main/java/com/ataulm/artcollector/artist/data/AndroidArtistRepository.kt
@@ -6,8 +6,8 @@ import com.ataulm.artcollector.ApiPersonRecord
 import com.ataulm.artcollector.HarvardArtMuseumApi
 import com.ataulm.artcollector.artist.domain.ArtistId
 import com.ataulm.artcollector.artist.domain.ArtistRepository
-import com.ataulm.artcollector.artist.domain.Gallery
 import com.ataulm.artcollector.domain.Artist
+import com.ataulm.artcollector.domain.Gallery
 import com.ataulm.artcollector.domain.Painting
 import javax.inject.Inject
