+++ artist/src/main/java/com/ataulm/artcollector/artist/domain/ArtistRepository.kt
@@ -1,6 +1,7 @@
 package com.ataulm.artcollector.artist.domain
 
 import com.ataulm.artcollector.domain.Artist
+import com.ataulm.artcollector.domain.Gallery
 
 internal interface ArtistRepository {
