+++ artist/src/main/java/com/ataulm/artcollector/artist/domain/Models.kt
@@ -1,7 +1,3 @@
 package com.ataulm.artcollector.artist.domain
 
-import com.ataulm.artcollector.domain.Painting
-
-internal class Gallery(collection: Collection<Painting>) : ArrayList<Painting>(collection)
-
 internal data class ArtistId(val value: String)
