+++ artist/src/main/java/com/ataulm/artcollector/artist/ui/ArtistViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.ataulm.artcollector.Event
-import com.ataulm.artcollector.artist.domain.Gallery
 import com.ataulm.artcollector.artist.domain.GetArtistGalleryUseCase
 import com.ataulm.artcollector.artist.domain.GetArtistUseCase
 import com.ataulm.artcollector.domain.Artist
+import com.ataulm.artcollector.domain.Gallery
 import com.ataulm.artcollector.domain.Painting
 import kotlinx.coroutines.*
 import javax.inject.Inject
